Update patch set 1
Patch Set 1:

Hi Marco,

Whenever any media file is opened this function will get hit to check whether this is a midi content or not. Because of the fd leak here (for files  > 2GB), even after the said media content is deleted, free size of the disk will not increase. This is the issue.

Patch-set: 1